Add files via upload
Hey everyone,
I have uploaded the code file and some images of the solarized water filter that we created as part of our project-based learning. The key highlight of this project is that it operates on the principle of solar energy, and the sediment filters are made from natural resources, which overall reduces the production cost.